style(mod): moved ApiResponseBody struct to into_url module
