+++ src/implementation/types.rs
@@ -5,13 +5,6 @@ use crate::{
     Category, NekosLifeError,
 };
 
-// represents the response body
-// for serde deserialization
-#[derive(serde::Deserialize, Debug)]
-pub(crate) struct ApiResponseBody {
-    pub(crate) url: UrlString,
-}
-
 /// a trait that must be implemented to be passed by the [`get`] function.
 ///
 /// this is slightly different from the [`TryInto`] trait in `std`,
@@ -31,74 +24,4 @@ pub trait IntoUrl {
 
     /// parse the body of the response
     fn parse(res: reqwest::Response) -> Self::Fut;
-}
-
-impl IntoUrl for &'static str {
-    type Response = crate::types::UrlString;
-    type Fut = std::pin::Pin<
-        Box<
-            dyn std::future::Future<
-                Output = types::Result<Self::Response>,
-            >,
-        >,
-    >;
-
-    fn into_url(self) -> types::Result<url::Url> {
-        Ok(string_to_endpoint!(
-            Into::<&'static str>::into(
-                &<Category as std::str::FromStr>::from_str(
-                    self
-                )
-                .map_err(|error| {
-                    NekosLifeError::UnknownEndpoint {
-                        error,
-                        url: self.to_owned(),
-                    }
-                })?
-            )
-        ))
-    }
-
-    fn parse(res: reqwest::Response) -> Self::Fut {
-        Box::pin(async move {
-            Ok(res.json::<ApiResponseBody>().await?.url)
-        })
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    // use pretty_assertions::assert_eq;
-
-    use {
-        crate::{NekosLifeError, BASEURL},
-        pretty_assertions::assert_eq,
-    };
-
-    #[test]
-    fn string_to_url() -> Result<(), NekosLifeError> {
-        use super::IntoUrl;
-
-        Ok(assert_eq!(
-            "Waifu".into_url()?.as_str(),
-            format!(
-                "{baseurl}img/waifu",
-                baseurl = BASEURL.as_str()
-            )
-        ))
-    }
-
-    #[tokio::test]
-    async fn parse_test() -> Result<(), NekosLifeError> {
-        Ok(assert!(lazy_regex::regex_is_match!(
-            r"^https://cdn\.nekos\.life/neko/[\w_.]+$",
-            &<&str as super::IntoUrl>::parse(
-                reqwest::get(
-                    BASEURL.join("img/")?.join("neko")?
-                )
-                .await?
-            )
-            .await?
-        )))
-    }
 }
